+++ docs/CHANGES.md
@@ -2,9 +2,8 @@
 
 ## 2.0.4 (Unreleased)
 - #737: Bugfix. Translations weren't being applied. [jcbrand]
-- Room information as returned by the server is now stored on the room model.
-  For context, see: http://xmpp.org/extensions/xep-0045.html#disco-roominfo
-  [jcbrand]
+- Fetch room info and store it on the room model.
+  For context, see: http://xmpp.org/extensions/xep-0045.html#disco-roominfo [jcbrand]
 - Bugfix. Switching from bookmarks form to config form shows only the spinner. [jcbrand]
 - Bugfix. Other room occupants sometimes not shown when reloading the page. [jcbrand]
 - Bugfix. Due to changes in `converse-core` the controlbox wasn't aware anymore of